# p9: remove unnecessary allocations when using sync.Pool