@hugelgupf 2020-08-11
I'm porting your change at https://github.com/google/gvisor/pull/3582 as well.